requirements.txt
praw
Keras
matplotlib
pillow
pandas
mlxtend
tensorflow-gpu
progressbar2
streamlit